airflow/requirements.txt


urllib3==1.25.11
clickhouse-connect==0.6.23
pandas==1.3.4
apache-airflow[celery]==2.8.0 --constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.8.0/constraints-3.9.txt"
python-dotenv==0.19.2